+++ src/index.ts
@@ -2,3 +2,4 @@ import * as apollo from "./apollo"
 export { apollo }
 export { default as query } from "./query"
 export { default as mutation } from "./mutation"
+export * from "./types"