+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 
 ## [Unreleased]
+
+## [0.3.0] - 2018-03-30
 ### Added
 - Add hyperapp-apollo.js and hyperapp-apollo.min.js to dist files.
 - Enable to use `<Query/>` in a loop.
@@ -32,6 +34,7 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/yuku-t/hyperapp-apollo/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/yuku-t/hyperapp-apollo/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/yuku-t/hyperapp-apollo/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/yuku-t/hyperapp-apollo/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/yuku-t/hyperapp-apollo/compare/2134207...v0.1.0